Pre-Alpha 4.0.11 Expanding the Questing System
Added item quest type under questobjects. Checks for the amount of a given item player has and references that against amountrequired.

Also added text and animations for when you accept and complete a quest.

Added a text doc explaining how to add a test and explaining various bits .

+++ Assets/coinscript.cs
@@ -10,7 +10,7 @@ public class coinscript : MonoBehaviour
     [SerializeField] AudioClip Jingle;
 
     public float despawnTimer;
-    public float coinValue = 1;
+    public int coinValue = 1;
     public float attractRadius = 1;
 
     // Start is called before the first frame update
